DEV: eliminated src/internal directory to make things      more consistent with sv4gui
DEV: added Makefile
