+++ org.sv.gui.qt.hello/Makefile
@@ -1,12 +1,11 @@
 # Hey emacs, this is a -*- makefile -*-
 
-# Copyright (c) 2009-2011, Open Source Medical Software Corporation,
-#              University of California, San Diego.
+# Copyright (c) Stanford University, The Regents of the University of
+#               California, and others.
 #
-# All rights reserved.
+# All Rights Reserved.
 #
-# Portions copyright (c) 1999-2007 Stanford University,
-# Nathan Wilson, Ken Wang, Charles Taylor.
+# See Copyright-SimVascular.txt for additional details.
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -19,15 +18,26 @@
 # The above copyright notice and this permission notice shall be included
 # in all copies or substantial portions of the Software.
 #
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-# OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
-# CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-# TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+# IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+# TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+# PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+# OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+# PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+# LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+# NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-TOP=C:/cygwin64/home/nwilson/gitwork/simvascular/branches/2018-01-22-mac_svexternals/BuildWithMake
+# need to set to build!
+# SV_TOP: set environment variable "SV_TOP" to point to core SV build "BuildWithMake" directory!
+
+# need to set before running executable at command line
+# SV_PLUGIN_PATH: location for modules and shared objects (: separated linux, ; separated windows)
+# SV_CUSTOM_PLUGINS: location for qt plugins (: separated linux, ; separated windows)
+
+TOP=$(SV_TOP)
 
 include $(TOP)/include.mk
 
@@ -63,8 +73,8 @@ QRCFILES += org_sv_gui_qt_hello_manifest.qrc \
             org_sv_gui_qt_hello_cached.qrc
 
 RCFILES = \
-    ../../plugin.xml \
-    ../../resources/hello.png
+    plugin.xml \
+    resources/hello.png
 
 CXXSRCS += $(addprefix moc_,$(HDRS:.h=.cxx))
 
@@ -89,7 +99,7 @@ DLLLIBS +=  \
           $(SVLIBFLAG)$(SV_LIB_UTILS_NAME)$(LIBLINKEXT) \
           $(SVLIBFLAG)$(SV_LIB_GLOBALS_NAME)$(LIBLINKEXT)
 
-EXTRA_MOC_INCDIRS = $(MITK_PLUGIN_INCDIRS) -I../../../org.sv.gui.qt.datamanager/src/internal
+EXTRA_MOC_INCDIRS = $(MITK_PLUGIN_INCDIRS) -I../../../org.sv.gui.qt.datamanager
 
 TARGET_LIB_NAME = $(SV_PLUGIN_HELLO_NAME)
