+++ org.sv.gui.qt.hello/svHello.cxx
@@ -1,0 +1,72 @@
+/* Copyright (c) Stanford University, The Regents of the University of
+ *               California, and others.
+ *
+ * All Rights Reserved.
+ *
+ * See Copyright-SimVascular.txt for additional details.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject
+ * to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+ * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+ * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+ * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+ * OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include "svHello.h"
+#include "ui_svHello.h"
+
+const QString svHello::EXTENSION_ID = "org.sv.views.hello";
+
+svHello::svHello() :
+  ui(new Ui::svHello)
+{
+
+
+}
+
+svHello::~svHello(){
+  delete ui;
+}
+
+void svHello::CreateQtPartControl(QWidget *parent){
+  m_Parent=parent;
+  ui->setupUi(parent);
+
+  m_DisplayWidget=GetActiveStdMultiWidget();
+
+  if(m_DisplayWidget==NULL)
+  {
+      parent->setEnabled(false);
+      MITK_ERROR << "Plugin Hello Init Error: No QmitkStdMultiWidget!";
+      return;
+  }
+
+  getText();
+  connect(ui->helloPushButton, SIGNAL(clicked()), this, SLOT(printText()) );
+  connect(ui->helloLineEdit, SIGNAL(editingFinished()), this, SLOT(getText()) );
+}
+
+void svHello::getText(){
+  hello_str = ui->helloLineEdit->text();
+}
+
+void svHello::printText(){
+  std::cout << hello_str << "\n";
+}
